requirements_gpt4.txt
decord==0.6.0
fastparquet==2024.2.0
google-ai-generativelanguage==0.6.3
google-api-core==2.19.0
google-api-python-client==2.129.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-generativeai @ git+https://github.com/google-gemini/generative-ai-python@efead6bea6768f6f4a3d90d348647b0a54fe2435
googleapis-common-protos==1.63.0
huggingface-hub==0.20.3
imageio==2.34.1
ipykernel==6.29.4
ipython==8.12.3
matplotlib==3.7.4
matplotlib-inline==0.1.6
matplotlib-scalebar==0.8.1
numpy
openai==1.10.0
opencv-python==4.9.0.80
palettable==3.3.3
pandas==2.0.3
PyYAML==6.0.1
rich==13.9.4
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.13.2
seaborn-image==0.9.0
tiktoken==0.5.2
torch==1.12.0
torchvision==0.13.0
tqdm==4.66.1
